chore: separate custom progress bar component
